Move dropdown/embed creation to handlers
